+++ extendr-api/tests/io_tests.rs
@@ -3,7 +3,7 @@ use ::extendr_api::io::Load;
 #[test]
 fn test_save() {
     use extendr_api::{io::PstreamFormat, io::Save, test, Result, Robj};
-    test! {
+    with_r(|| {
         let mut w = Vec::new();
         Robj::from(1).to_writer(&mut w, PstreamFormat::AsciiFormat, 3, None)?;
         assert!(w[0] == b'A');
@@ -16,13 +16,13 @@ fn test_save() {
         // Robj::from(1).save(&path, PstreamFormat::AsciiFormat, 3, None)?;
         // let s = std::fs::read(path).unwrap();
         // assert!(s.starts_with(b"A\n"));
-    }
+    });
 }
 
 #[test]
 fn test_load() {
     use extendr_api::{io::PstreamFormat, test, Result, Robj};
-    test! {
+    with_r(|| {
         let text = r#"A
 3
 262402
@@ -42,5 +42,5 @@ UTF-8
 
         let res = Robj::from_reader(&mut c, PstreamFormat::AsciiFormat, None);
         assert_eq!(res, Ok(Robj::from(1_i32)));
-    }
+    });
 }
