+++ extendr-api/tests/optional/either/either_mock_try_from_robj.rs
@@ -24,7 +24,7 @@ impl TryFrom<&Robj> for Failure {
 
 #[test]
 fn try_from_one_match() {
-    test! {
+    with_r(|| {
         let robj = r!(());
 
         let left_match = <Either<Success, Failure> as TryFrom<&Robj>>::try_from(&robj);
@@ -32,23 +32,23 @@ fn try_from_one_match() {
 
         assert_eq!(left_match, Ok(Left(Success{})));
         assert_eq!(right_match, Ok(Right(Success{})));
-    }
+    });
 }
 
 #[test]
 fn try_from_both_match_return_left() {
-    test! {
+    with_r(|| {
         let robj = r!(());
 
         let both_match = <Either<Success, Success> as TryFrom<&Robj>>::try_from(&robj);
 
         assert_eq!(both_match, Ok(Left(Success{})));
-    }
+    });
 }
 
 #[test]
 fn try_from_none_match_return_error() {
-    test! {
+    with_r(|| {
         let robj = r!(());
 
         let none_match = <Either<Failure, Failure> as TryFrom<&Robj>>::try_from(&robj);
@@ -61,5 +61,5 @@ fn try_from_none_match_return_error() {
                     Box::new(Error::Other(String::new())))
             )
         );
-    }
+    });
 }
